Scalable attributes balanced and defined for towers. Had to start implementing tower behaviors to have better understanding of how the scalable attributes would work. Fixed Exponential growth

+++ Assets/Scripts/TowerDefense/Game/GameScore.cs
@@ -7,7 +7,7 @@ namespace TowerDefense.Game
     /// </summary>
     public class GameScore: SerializableData
     {
-        public long score;
+        public float score;
         public int killCount;
         public float timeElapsed;
         public float damageDone;
